'CMakeLists.txt' (JiveXML-02-02-90)
   * tag JiveXML-02-02-90
   * finally fixed NICOS 'Major Warning' due to using
     AlgTool instead of AthAlgTool

2015-06-09  Jochen Meyer <jomeyer -at- mail.cern.ch>
            Juergen Thomas <jpt -at- hep.ph.bham.ac.uk>
   * tag JiveXML-02-02-89
   * addressing coverity defect (FormatTool)

2015-04-25  Juergen Thomas <jpt -at- hep.ph.bham.ac.uk>

   * tag JiveXML-02-02-88
   * change variable type for event number to 'unsigned long'
     from 'unsigned int', as high-rate tests can have
     > 4 billion as event number

...
(Long ChangeLog diff - truncated)

+++ graphics/JiveXML/JiveXML/EventStream.h
@@ -21,12 +21,12 @@ namespace JiveXML {
     public:
 
       //Constructor with all parameters
-      EventStreamID( unsigned int EventNumber, unsigned int RunNumber, std::string StreamName ) :
+      EventStreamID( unsigned long EventNumber, unsigned int RunNumber, std::string StreamName ) :
         m_event(EventNumber), m_run(RunNumber), m_stream(StreamName){
       };
 
       //Constructor with all paramters and char *
-      EventStreamID( unsigned int EventNumber, unsigned int RunNumber, const char* StreamName ) :
+      EventStreamID( unsigned long EventNumber, unsigned int RunNumber, const char* StreamName ) :
         m_event(EventNumber), m_run(RunNumber){
           //Check for NULL pointer
           if (StreamName) m_stream=StreamName;
@@ -38,7 +38,7 @@ namespace JiveXML {
       };
 
       //Get the different values
-      unsigned int EventNumber() const { return m_event; };
+      unsigned long EventNumber() const { return m_event; };
       unsigned int RunNumber() const { return m_run; };
       std::string StreamName() const { return m_stream; };
